refactor: split test_xml2_pipeline.py into smaller files (#24)
Split the large test file (2413 lines) into 5 focused test files:
- test_xml2_pipeline_args.py (296 lines) - argument parsing tests
- test_xml2_pipeline_processing.py (806 lines) - content processing tests
- test_xml2_pipeline_output.py (602 lines) - output file generation tests
- test_xml2_pipeline_cleaned_text.py (460 lines) - cleaned text file tests
- test_xml2_pipeline_integration.py (337 lines) - integration tests

All 79 tests preserved. Each file now under 1000 line limit.
